YEL-45 [chore] 코드 정리
