--- src/main/java/com/yello/server/domain/cooldown/response/VoteListResponse.java
@@ -1,18 +1,0 @@
-package com.yello.server.domain.cooldown.response;
-
-import java.util.List;
-import lombok.Builder;
-
-@Builder
-public record VoteListResponse(
-    Integer totalCount,
-    List<VoteResponse> votes
-) {
- 
-    public static VoteListResponse of(Integer totalCount, List<VoteResponse> votes) {
-        return VoteListResponse.builder()
-            .totalCount(totalCount)
-            .votes(votes)
-            .build();
-    }
-}
