+++ src/main/java/com/yello/server/domain/vote/dto/response/VoteListResponse.java
@@ -1,0 +1,19 @@
+package com.yello.server.domain.vote.dto.response;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record VoteListResponse(
+        Integer totalCount,
+        List<VoteResponse> votes
+) {
+
+    public static VoteListResponse of(Integer totalCount, List<VoteResponse> votes) {
+        return VoteListResponse.builder()
+                .totalCount(totalCount)
+                .votes(votes)
+                .build();
+    }
+}
