+++ src/main/java/com/yello/server/domain/vote/service/VoteService.java
@@ -2,15 +2,16 @@ package com.yello.server.domain.vote.service;
 
 import com.yello.server.domain.cooldown.response.KeywordCheckResponse;
 import com.yello.server.domain.cooldown.response.RevealNameResponse;
-import com.yello.server.domain.cooldown.response.VoteCreateResponse;
-import com.yello.server.domain.cooldown.response.VoteDetailResponse;
-import com.yello.server.domain.cooldown.response.VoteFriendResponse;
-import com.yello.server.domain.cooldown.response.VoteListResponse;
 import com.yello.server.domain.question.dto.response.VoteAvailableResponse;
 import com.yello.server.domain.question.dto.response.VoteQuestionResponse;
 import com.yello.server.domain.vote.dto.request.CreateVoteRequest;
-import java.util.List;
+import com.yello.server.domain.vote.dto.response.VoteCreateResponse;
+import com.yello.server.domain.vote.dto.response.VoteDetailResponse;
+import com.yello.server.domain.vote.dto.response.VoteFriendResponse;
+import com.yello.server.domain.vote.dto.response.VoteListResponse;
 import org.springframework.data.domain.Pageable;
+
+import java.util.List;
 
 public interface VoteService {
