+++ src/main/java/com/yello/server/domain/vote/service/VoteServiceImpl.java
@@ -1,28 +1,9 @@
 package com.yello.server.domain.vote.service;
 
-import static com.yello.server.domain.vote.common.WeightedRandom.randomPoint;
-import static com.yello.server.global.common.ErrorCode.INVALID_VOTE_EXCEPTION;
-import static com.yello.server.global.common.ErrorCode.LACK_POINT_EXCEPTION;
-import static com.yello.server.global.common.ErrorCode.LACK_USER_EXCEPTION;
-import static com.yello.server.global.common.ErrorCode.NOT_FOUND_VOTE_EXCEPTION;
-import static com.yello.server.global.common.ErrorCode.USERID_NOT_FOUND_USER_EXCEPTION;
-import static com.yello.server.global.common.util.ConstantUtil.KEYWORD_HINT_POINT;
-import static com.yello.server.global.common.util.ConstantUtil.NAME_HINT_DEFAULT;
-import static com.yello.server.global.common.util.ConstantUtil.NAME_HINT_POINT;
-import static com.yello.server.global.common.util.ConstantUtil.RANDOM_COUNT;
-import static com.yello.server.global.common.util.ConstantUtil.TIMER_FIFTY_TIME;
-import static com.yello.server.global.common.util.ConstantUtil.VOTE_COUNT;
-import static com.yello.server.global.common.util.TimeUtil.minusTime;
-
 import com.yello.server.domain.cooldown.entity.Cooldown;
 import com.yello.server.domain.cooldown.entity.CooldownRepository;
 import com.yello.server.domain.cooldown.response.KeywordCheckResponse;
 import com.yello.server.domain.cooldown.response.RevealNameResponse;
-import com.yello.server.domain.cooldown.response.VoteCreateResponse;
-import com.yello.server.domain.cooldown.response.VoteDetailResponse;
-import com.yello.server.domain.cooldown.response.VoteFriendResponse;
-import com.yello.server.domain.cooldown.response.VoteListResponse;
-import com.yello.server.domain.cooldown.response.VoteResponse;
 import com.yello.server.domain.friend.entity.Friend;
 import com.yello.server.domain.friend.entity.FriendRepository;
 import com.yello.server.domain.friend.exception.FriendException;
@@ -38,20 +19,27 @@ import com.yello.server.domain.user.entity.User;
 import com.yello.server.domain.user.entity.UserRepository;
 import com.yello.server.domain.user.exception.UserException;
 import com.yello.server.domain.vote.dto.request.CreateVoteRequest;
+import com.yello.server.domain.vote.dto.response.*;
 import com.yello.server.domain.vote.entity.Vote;
 import com.yello.server.domain.vote.entity.VoteRepository;
 import com.yello.server.domain.vote.exception.VoteForbiddenException;
 import com.yello.server.domain.vote.exception.VoteNotFoundException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
+import static com.yello.server.domain.vote.common.WeightedRandom.randomPoint;
+import static com.yello.server.global.common.ErrorCode.*;
+import static com.yello.server.global.common.util.ConstantUtil.*;
+import static com.yello.server.global.common.util.TimeUtil.minusTime;
 
 @Service
 @RequiredArgsConstructor
@@ -68,9 +56,9 @@ public class VoteServiceImpl implements VoteService {
     public VoteListResponse findAllVotes(Long userId, Pageable pageable) {
         Integer count = voteRepository.countAllByReceiverUserId(userId);
         List<VoteResponse> votes = voteRepository.findAllByReceiverUserId(userId, pageable)
-            .stream()
-            .map(VoteResponse::of)
-            .toList();
+                .stream()
+                .map(VoteResponse::of)
+                .toList();
         return VoteListResponse.of(count, votes);
     }
 
@@ -84,7 +72,7 @@ public class VoteServiceImpl implements VoteService {
 
     @Override
     public List<VoteFriendResponse> findAllFriendVotes(Long userId, Pageable pageable) {
-        //todo 후순위
+        voteRepository.find
         return null;
     }
 
@@ -92,7 +80,7 @@ public class VoteServiceImpl implements VoteService {
     @Override
     public KeywordCheckResponse checkKeyword(Long userId, Long voteId) {
         Vote vote = voteRepository.findById(voteId)
-            .orElseThrow(() -> new VoteNotFoundException(NOT_FOUND_VOTE_EXCEPTION));
+                .orElseThrow(() -> new VoteNotFoundException(NOT_FOUND_VOTE_EXCEPTION));
 
         User user = findUser(userId);
         vote.updateKeywordCheck();
@@ -120,8 +108,8 @@ public class VoteServiceImpl implements VoteService {
         Collections.shuffle(question);
 
         List<Question> yelloQuestionList = question.stream()
-            .limit(VOTE_COUNT)
-            .toList();
+                .limit(VOTE_COUNT)
+                .toList();
 
         yelloQuestionList.forEach(yello -> yelloVoteList.add(getVoteData(user, yello)));
 
@@ -138,7 +126,7 @@ public class VoteServiceImpl implements VoteService {
         }
 
         Cooldown cooldown = cooldownRepository.findByUserId(user.getId())
-            .orElse(Cooldown.of(user, minusTime(LocalDateTime.now(), TIMER_FIFTY_TIME)));
+                .orElse(Cooldown.of(user, minusTime(LocalDateTime.now(), TIMER_FIFTY_TIME)));
 
         return VoteAvailableResponse.of(user, cooldown);
     }
@@ -150,13 +138,13 @@ public class VoteServiceImpl implements VoteService {
         sender.plusPoint(request.totalPoint());
 
         request.voteAnswerList().forEach(vote ->
-            voteRepository.save(
-                Vote.createVote(
-                    vote.keywordName(),
-                    sender,
-                    findUser(vote.friendId()),
-                    questionService.findByQuestionId(vote.questionId()), vote.colorIndex())
-            )
+                voteRepository.save(
+                        Vote.createVote(
+                                vote.keywordName(),
+                                sender,
+                                findUser(vote.friendId()),
+                                questionService.findByQuestionId(vote.questionId()), vote.colorIndex())
+                )
         );
 
         Optional<Cooldown> cooldown = cooldownRepository.findByUserId(sender.getId());
@@ -176,7 +164,7 @@ public class VoteServiceImpl implements VoteService {
         Vote vote = findVote(voteId);
         String name = vote.getSender().getName();
 
-        if (vote.getNameHint()!=NAME_HINT_DEFAULT) {
+        if (vote.getNameHint() != NAME_HINT_DEFAULT) {
             throw new VoteNotFoundException(INVALID_VOTE_EXCEPTION);
         }
         if (sender.getPoint() < NAME_HINT_POINT) {
@@ -188,9 +176,9 @@ public class VoteServiceImpl implements VoteService {
         sender.minusPoint(NAME_HINT_POINT);
 
         return RevealNameResponse.builder()
-            .name(name.charAt(randomIndex))
-            .nameIndex(randomIndex)
-            .build();
+                .name(name.charAt(randomIndex))
+                .nameIndex(randomIndex)
+                .build();
     }
 
     public VoteQuestionResponse getVoteData(User user, Question question) {
@@ -198,11 +186,11 @@ public class VoteServiceImpl implements VoteService {
         Collections.shuffle(keywordList);
 
         return VoteQuestionResponse.builder()
-            .friendList(getFriendList(user))
-            .keywordList(getKeywordList(question))
-            .question(VoteContentVO.of(question))
-            .questionPoint(randomPoint())
-            .build();
+                .friendList(getFriendList(user))
+                .keywordList(getKeywordList(question))
+                .question(VoteContentVO.of(question))
+                .questionPoint(randomPoint())
+                .build();
     }
 
     public List<VoteShuffleFriend> getFriendList(User user) {
@@ -210,9 +198,9 @@ public class VoteServiceImpl implements VoteService {
         Collections.shuffle(allFriend);
 
         return allFriend.stream()
-            .map(VoteShuffleFriend::of)
-            .limit(RANDOM_COUNT)
-            .collect(Collectors.toList());
+                .map(VoteShuffleFriend::of)
+                .limit(RANDOM_COUNT)
+                .collect(Collectors.toList());
     }
 
     public List<String> getKeywordList(Question question) {
@@ -220,19 +208,19 @@ public class VoteServiceImpl implements VoteService {
         Collections.shuffle(keywordList);
 
         return keywordList.stream()
-            .map(Keyword::getKeywordName)
-            .limit(RANDOM_COUNT)
-            .toList();
+                .map(Keyword::getKeywordName)
+                .limit(RANDOM_COUNT)
+                .toList();
     }
 
     private Vote findVote(Long id) {
         return voteRepository.findById(id)
-            .orElseThrow(() -> new VoteNotFoundException(NOT_FOUND_VOTE_EXCEPTION));
+                .orElseThrow(() -> new VoteNotFoundException(NOT_FOUND_VOTE_EXCEPTION));
     }
 
     private User findUser(Long userId) {
         return userRepository.findById(userId)
-            .orElseThrow(() -> new UserException(USERID_NOT_FOUND_USER_EXCEPTION));
+                .orElseThrow(() -> new UserException(USERID_NOT_FOUND_USER_EXCEPTION));
     }
 
 }
